so i had absolutely no idea what i was doing last commit
