+++ README.md
@@ -85,3 +85,4 @@ To syscall to the kernel, use `int 0x33`. The service code goes in `ah`, and the
 | `15` | Program loading error, e.g. not a binary file |
 | `18` | Error in configuration |
 | `23` | Buffer space exceeded (Command too long) |
+| `28` | Error in allocating handle (app handle, file handle, etc) |
