+++ kernel/kapps/shell.hh
@@ -1,4 +1,22 @@
-#include "sh_builtins.hh"
+// execute program
+extern "C" int prog(int arg);
+
+const char* comnames[] = {
+  "clr",
+  "\xff\xf0 shift+f10",
+  "exec",
+  "\xff<u32>",
+  "fread", 
+  "fstat",
+  "\xff[inode|filename]",
+  "help",
+  "ls",
+  "reset",
+  "sys"
+  ":[conf|cpu|disk|indic]",
+  "time",
+  NULL
+};
 
 // shell applet
 #define SHELL_PROMPT_WIDTH 4
@@ -86,35 +104,171 @@ private: // hidden fields (only for internal use)
 
   void shexec() {
     if (strlen(this->work.buf) == 0) return;
+    outbuf = malloc(OUTBUF_LEN);
 
-    char found = -1; // have we found a command yet
-    for (int cm = 0; cmdlist[cm].name; ++cm) {
-      int checked_len = strlen(cmdlist[cm].name);
-      if (memcmp(this->work.buf, cmdlist[cm].name, checked_len) && is_whitespace(this->work.buf[checked_len])) { // memcmp, because of arguments, and check that command doesnt go on
-        found = cm;
-        break;
+    unsigned char fmt = 0;
+
+    int wher; // first whitespace character
+    for (wher = 0; !is_whitespace(work.buf[wher]); ++wher);
+
+    char *args = &(work.buf[wher + 1]);
+    char safe_padding = work.buf[wher]; // keep the padding character, so that history works
+    work.buf[wher] = '\0'; // make the command the only relevant word
+
+    if (strcmp(work.buf, "clr")) { // clear screen
+      clear_scr();
+      set_cur(0);
+    } else if (strcmp(work.buf, "exec")) {
+      if (disk_config -> qi_magic != CONFIG_MAGIC) {
+        msg(KERNERR, E_NOSTORAGE, "Disk is unavailable");
+        goto shell_clean;
       }
-    }
 
-    if (found == -1) {
+      read_inode(
+        name2inode("prog.bin"),
+        0x100000
+      ); // reads disk, has to get master or slave
+
+      int ar = -1;
+      if (strlen(args)) {
+        ar = to_uint(args);
+      }
+
+      enum ERRSIG ret = prog(ar);
+      if (ret == E_BOUND) {
+        msg(PROGERR, ret, "Program not found");
+      } else if (ret == E_ILLEGAL) {
+        msg(KERNERR, ret, "Program executed illegal instruction");
+      } else {
+        line_feed();
+      }
+    } else if (strcmp(work.buf, "fread")) {
+      char *datablk = malloc(disk_config -> wdata.len << 9);
+      read_inode(
+        name2inode("data.txt"),
+        datablk
+      ); // reads disk, has to get master or slave
+      write_str(datablk, COLOUR(BLACK, WHITE));
+
+      free(datablk);
+      line_feed();
+    } else if (strcmp(work.buf, "fstat")) {
+      int ar = -1;
+      if (strlen(args)) {
+        ar = to_uint(args);
+        if (ar < 0) {
+          ar = name2inode(args);
+          if (ar < 0) goto shell_clean;
+        }
+      }
+      
+      if (ar < 0 || !(file_table[ar].name)) {
+        msg(PROGERR, E_NOFILE, "Invalid inode");
+        goto shell_clean;
+      }
+
+      sprintf(outbuf, "%s <%d>\n\t%d bytes, sector %2X\n\tModified %4d-%2d-%2d %2d:%2d:%2d",
+        file_table[ar].name,
+        ar,
+        file_table[ar].loc.len << 9,
+        &(file_table[ar].loc.lba),
+        file_table[ar].modified.year,
+        file_table[ar].modified.month,
+        file_table[ar].modified.date,
+        file_table[ar].modified.hour,
+        file_table[ar].modified.minute,
+        file_table[ar].modified.second
+      );
+
+      fmt = COLOUR(GREEN, RED);
+    } else if (strcmp(work.buf, "help")) {
+      for (int cm = 0; comnames[cm]; ++cm) {
+        if (comnames[cm][0] == -1) {
+          sprintf(endof(outbuf), " %s", comnames[cm] + 1);
+        } else if (comnames[cm][0] == ':') {
+          strcat(outbuf, comnames[cm]);
+        } else {
+          sprintf(endof(outbuf), "%c\t%s", cm ? '\n' : 0, comnames[cm]);
+        }
+      }
+
+      fmt = COLOUR(BLUE, B_MAGENTA);
+    } else if (strcmp(work.buf, "ls")) {
+      for (int filek = 0; file_table[filek].name; filek++) {
+        strcat(outbuf, file_table[filek].name);
+        *endof(outbuf) = '\t';
+      }
+
+      fmt = COLOUR(BLACK, B_WHITE);
+    } else if (strcmp(work.buf, "reset")) {
+      cpu_reset();
+      // if you reach past here, something has gone very wrong indeed
+    } else if (strcmp(work.buf, "sys:cpu")) {
+      sprintf(outbuf, "CPUID.\n\t\x10 %12s\n\tFamily %2xh, Model %2xh, Stepping %1xh\n\tBrand \"%s\"",
+        &(hardware -> vendor),
+        &(hardware -> c_family), // family
+        &(hardware -> c_model), // model
+        &(hardware -> c_stepping), // stepping
+        hardware -> cpuid_ext_leaves >= 0x80000004 ? &(hardware -> brand) : NULL // brand string
+      );
+
+      fmt = COLOUR(YELLOW, B_GREEN); // fmt
+    } else if (strcmp(work.buf, "sys:conf")) {
+      sprintf(outbuf, "config.qi\n\tProgram at lba sector %2X, %d sector(s)\n\t\x10\t%s\n\tWritable data at lba sector %2X, %d sector(s)",
+        &(disk_config -> exec.lba),
+        disk_config -> exec.len,
+        hardware -> boot_disk_p.itrf_type,
+        &(disk_config -> wdata.lba),
+        disk_config -> wdata.len
+      );
+
+      fmt = COLOUR(BLUE, B_YELLOW); // fmt
+    } else if (strcmp(work.buf, "sys:disk")) {
+      sprintf(outbuf, "CSDFS Disk:\n\tVol. label \"%16s\"\n\tVol. ID %16X\n\tDisk %2xh\n\tVolume size %d",
+        &(csdfs -> label),
+        &(csdfs -> vol_id),
+        &(hardware -> bios_disk),
+        csdfs -> fs_size * SECTOR_LEN
+      );
+
+      fmt = COLOUR(RED, B_YELLOW); // fmt
+    } else if (strcmp(work.buf, "sys:indic")) {
+      // indicators
+      sprintf(outbuf, "scroll: %d\nnum: %d\ncaps: %d",
+        bittest(&(keys -> modifs), 0),
+        bittest(&(keys -> modifs), 1),
+        bittest(&(keys -> modifs), 2)
+      );
+
+      fmt = COLOUR(GREEN, RED);
+    } else if (strcmp(work.buf, "time")) {
+      sprintf(outbuf, "Time since kernel load: %d.%2ds\n%s%c%4d-%2d-%2d %2d:%2d:%2d",
+        countx / 100,
+        countx % 100,
+        curr_time -> weekday ? weekmap[curr_time -> weekday - 1] : NULL,
+        curr_time -> weekday ? ' ' : 0,
+        curr_time -> year,
+        curr_time -> month,
+        curr_time -> date,
+        curr_time -> hour,
+        curr_time -> minute,
+        curr_time -> second
+      );
+
+      fmt = COLOUR(RED, B_CYAN);
+    } else {
       msg(WARN, E_UNKENTITY, "Unknown command");
-      return;
+      goto shell_clean;
     }
 
-    char *outbuf = malloc(OUTBUF_LEN);
-
-    // we pass the out buffer, and the arguments
-    unsigned char result = (*cmdlist[found].func)(
-      outbuf,
-      this->work.buf + strlen(cmdlist[found].name) + 1
-    );
-
-    if (result) {
-      write_str(outbuf, result);
+    if (fmt) {
+      write_str(outbuf, fmt);
       line_feed();
     }
 
+  shell_clean:
     free(outbuf);
+    work.buf[wher] = safe_padding; // ok so jank way of keeping history intact
   }
 
 public:
