+++ kernel/misc.c
@@ -5,7 +5,7 @@ struct far_ver * curr_ver = &((struct far_ver) {
   .major = 0,
   .minor = 0,
   .patch = 3,
-  .build = 101, // minor changes + hotfixes
+  .build = 102, // minor changes + hotfixes
 });
 
 struct farb_header __seg_fs *prog_head = 0;
